requirements.txt
Flask>=3.0
PyMySQL>=1.0
Flask-Session>=0.8.0
flask-cors>=3.0
python-dotenv>=0.21
